Only set MouseEvent getters if they are configurable
The .pageX and .pageY properties of PhantomJS have their configurable property
set to false.  When trying to simulate mouse events in PhantomJS where the
page{X|Y} values are set to 0, calls to defineProperty throws the following
exception:

TypeError: Attempting to change access mechanism for an unconfigurable property.

Checking whether the property is configurable further tightens the number of
possible cases for triggering this 'IE quirk' allowing event simulation in
PhantomJS to work.